requirements.txt
mcp-server>=0.1.0
ipaddress>=1.0.23
    